+++ test/th5s.c
@@ -70,7 +70,7 @@ struct space4_struct {
     unsigned u;
     float f;
     char c2;
- } space4_data={'v',987123,-3.14,'g'}; /* Test data for 4th dataspace */
+ } space4_data={'v',987123,(float)-3.14,'g'}; /* Test data for 4th dataspace */
 
 /****************************************************************
 **
@@ -509,7 +509,7 @@ test_h5s_chunk(void)
     /* Initialize float array */
     for(i=0; i<50000; i++)
         for(j=0; j<3; j++)
-            chunk_data_flt[i][j]=i*2.5-j*100.3;
+            chunk_data_flt[i][j]=(float)(i*2.5-j*100.3);
 
     status= H5Dwrite(dsetID,H5T_NATIVE_FLOAT,H5S_ALL,H5S_ALL,H5P_DEFAULT,chunk_data_flt);
     CHECK(status, FAIL, "H5Dwrite");
